EXTENSION_SERVER/requirements.txt
Flask
xmltodict
Gunicorn
Werkzeug==2.2.2
python-dotenv
Flask-Cors
Flask-SQLAlchemy==2.5.0
SQLAlchemy==1.3.24
requests
psycopg2
Flask-Migrate==4.0.4
pandas
watchdog
sentry-sdk[flask]
urllib3==1.26.6
tabulate
python-telegram-bot
telegram
flask-swagger-ui
flask-restplus
pydot
graphviz
matplotlib
networkx
flasgger
flask_monitoringdashboard
alive_progress
tqdm
bs4
tinydb
beautifulsoup4
lxml